add AbstractService.java     ServiceDto.java     UsersService.java modify UsersDto.java        UsersRepository.java

+++ src/main/java/it/classe/SpringClass/Repository/UsersRepository.java
@@ -6,8 +6,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface UsersRepository extends JpaRepository<Users, Long> {
+public interface UsersRepository extends JpaRepository<Users, Integer> {
 
     Users findByUsername(String username);
